@@ -1,9 +1,8 @@
-using Tavis.Models;
+using TavisApi.V2.Users;
 
 namespace TavisApi.Services;
 
-public interface IUserService
-{
-  public string? GetCurrentUserName();
-  public User? GetCurrentUser();
+public interface IUserService {
+	public string? GetCurrentUserName();
+	public User? GetCurrentUser();
 }
